list/eol.py tweaks
* Move Position[] from grab-bag List (misc) functions here
* order imports
* make explicit Context in Symbol() creation
* add TODO list of builtins needed to fill out subsection more completely
* move closer towards A_ attribute flag naming
